Create Mirror reflection Problem
Reflections Problem (Tailored from Project Euler site):
(10 points)
In laser physics, a "white cell"  is a mirror system that acts as a delay line for the laser
beam. The beam enters the cell, bounces around on the mirrors, and eventually works
its way back out.
The specific white cell we will be considering is a circle with the equation x 2  + y 2  = 100
The section corresponding to −0.01 ≤ x ≤ +0.01 at the top is missing
The light beam in this problem starts at the
point (1.0,9.95) just outside the white cell, and the beam has a slope of first impacts the
mirror at (2,-6).
Each time the laser beam hits the surface of the circle, it follows the usual law of
reflection &quot;angle of incidence equals angle of reflection.&quot; That is, both the incident and
reflected beams make the same angle with the normal line at the point of incidence.
In the figure on the left, the red line shows the first two points of contact between the
laser beam and the wall of the white cell; the blue line shows the line tangent to the
circle at the point of incidence of the first bounce.
The slope m of the tangent line at any point (x,y) of the given circle is: m = −x/y
The normal line is perpendicular to this tangent line at the point of incidence.
The animation on the right shows the first 10 reflections of the beam.
How many times does the beam hit the internal surface of the white cell before exiting?